fix(warn): Silence C4267 on 64bit builds.

+++ Code/Commando/CDKeyAuth.cpp
@@ -149,7 +149,7 @@ void CCDKeyAuth::AuthSerial(const char *challenge, StringClass &resp) {
 	*outb = 0;
 
 	// MD5 Hash Here.
-	MD5Digest((BYTE *)cdkey, strlen(cdkey), md5hash);
+	MD5Digest((BYTE *)cdkey, unsigned(strlen(cdkey)), md5hash);
 
 	// hashserial, challenge, outbuf
 	gcd_compute_response(cdkey, (char *)challenge, response, CDResponseMethod_NEWAUTH);
